refactor(boylove): merge `parser.rs` into `lib.rs`
